changed resource count type to long int

+++ include/MaterialManager.hpp
@@ -34,7 +34,7 @@ public:
 	struct Resource
 	{
 		std::string name;
-		int count;
+		long count;
 	};
 
 	/**
@@ -141,7 +141,7 @@ public:
 	 * @return const std::unordered_map<std::string, int>& A reference to the
 	 * internal resources.
 	 */
-	const std::map<std::string, int> &getResources();
+	const std::map<std::string, long> &getResources();
 
 	/**
 	 * @brief Get a pointer to the icon texture for the specified resource.
@@ -156,7 +156,7 @@ private:
 	 * @brief Internal map of the name to the count of each resource.
 	 *
 	 */
-	std::map<std::string, int> mResources;
+	std::map<std::string, long> mResources;
 
 	/**
 	 * @brief Map of resources to their texture.
